Completed conversion to Rails. Todo: Resource routes

+++ app/controllers/posts_controller.rb
@@ -3,7 +3,18 @@ class PostsController < ApplicationController
   end
 
   def edit
-    @post = Post.find(params[:id])
+    if session[:user]
+      @post = Post.find(params[:id])
+      if session[:user]["id"] == @post.user_id || session[:user]["admin"]
+        render :edit
+      else
+        @message = "Forbidden action"
+        render "layouts/message"
+      end
+    else
+      @message = "Forbidden action"
+      render "layouts/message"
+    end
   end
 
   def create
@@ -19,10 +30,10 @@ class PostsController < ApplicationController
 
   def delete
     post = Post.find(params[:id])
-    if session[:user]["admin"] == 1 && post
+    if session[:user] && session[:user]["admin"] == 1 && post
       post.destroy
       User.where(id: post.user_id).update_all('total_posts = total_posts - 1')
-      redirect_to '/threads/#{post.mb_thread.}'
+      redirect_to thread_path(params[:board_id], params[:thread_id])
     else
       @message = "Error: Forbidden action"
       render 'layouts/message'
@@ -30,10 +41,10 @@ class PostsController < ApplicationController
   end
 
   def update
-    post = Post.find(params[:post][:post_id])
+    post = Post.find(params[:id])
     post.message = params[:post][:message]
     post.save
-    redirect_to thread_path(post.mb_thread_id)
+    redirect_to thread_path(params[:board_id], params[:thread_id])
   end
 
 end
